requirements.txt
# Microsoft Graph API y authentication 
msal==1.30.0

# Procesamiento de datos
pandas==2.3.1
openpyxl==3.1.5

# Configuración y variables de entorno
python-dotenv==1.0.0

# HTTP requests
requests==2.32.4

# Dependencias que ya tienes instaladas
certifi==2025.7.14
charset-normalizer==3.4.2
et-xmlfile==2.0.0
idna==3.10
numpy==2.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
urllib3==2.5.0

# Utilidades adicionales (opcional)
colorama==0.4.6
Flask==3.0.0
